mp2/requirements.txt
numpy
scipy
tqdm
opencv-python
matplotlib
torch
torchvision
tensorboard
pykitti
rerun-sdk
shapely
